+++ photoscrawl/internal/photos/errors.go
@@ -1,0 +1,5 @@
+package photos
+
+import "errors"
+
+var ErrExportAlreadyRunning = errors.New("photokit export already running")